requirements.txt
Django>=5.2.3
python-telegram-bot>=22.1
Pillow>=11.2.1
python-dotenv>=1.1.0
djangorestframework>=3.16.0
django-widget-tweaks>=1.5.0 